Throw exception upon failure instead of arbitrary default value (#21)
The reasoning is that depending of the context, one may want to apply a different arbitrary default value. It is easier to do that by wrapping the call in a try/catch catching a specific exception rather than getting the value directly (since you wouldn't know if you get the right value found, or the default one).