feat: Add a new method to get the number of CPU cores to use for parallelism (#127)
I think the current primitives offered by the library are fine. What is not is that people use `$counter->getCount()` with the expectation that it is the number of cores they should use for parallelism (and it is the usage I've defaulted to as well, as this is how it was used in the original code that this library was taken from). But as highlighted in #125, it is not. A non exhaustive list:

- It does not account for the fact that the current process reserves one core.
- It does not account for a user defined limit (e.g. like mentioned in #125 with Kubernetes).
- It does not account for the current system load.

Seeing [`uv_available_parallelism`](https://docs.libuv.org/en/v1.x/misc.html#c.uv_available_parallelism), I've decide to introduce a new API that can address those points and make it easier for consumers to get the appropriate number of cores for the purpose of knowing the number of processes they can launch for parallelism, which is the main usage of this library.

I will add more parameters in the future to address the points mentioned above, but they will be added as default parameters hence will not break the API.